move `Uint8Array` to / from base64 and hex to stable ES
