+++ packages/core-js-compat/src/modules-by-versions.mjs
@@ -302,4 +302,12 @@ export default {
   3.44: [
     'esnext.iterator.sliding',
   ],
+  3.45: [
+    'es.uint8-array.from-base64',
+    'es.uint8-array.from-hex',
+    'es.uint8-array.set-from-base64',
+    'es.uint8-array.set-from-hex',
+    'es.uint8-array.to-base64',
+    'es.uint8-array.to-hex',
+  ],
 };
