+++ packages/core-js-pure/override/modules/es.uint8-array.set-from-base64.js
@@ -1,0 +1,1 @@
+// empty
