+++ packages/core-js-pure/override/modules/es.uint8-array.to-base64.js
@@ -1,0 +1,1 @@
+// empty
