+++ packages/core-js-pure/override/modules/es.uint8-array.to-hex.js
@@ -1,0 +1,1 @@
+// empty
