+++ packages/core-js/actual/typed-array/from-base64.js
@@ -1,2 +1,5 @@
 'use strict';
+var parent = require('../../stable/typed-array/from-base64');
 require('../../modules/esnext.uint8-array.from-base64');
+
+module.exports = parent;
