+++ packages/core-js/actual/typed-array/from-hex.js
@@ -1,2 +1,5 @@
 'use strict';
+var parent = require('../../stable/typed-array/from-hex');
 require('../../modules/esnext.uint8-array.from-hex');
+
+module.exports = parent;
