+++ packages/core-js/actual/typed-array/set-from-base64.js
@@ -1,2 +1,5 @@
 'use strict';
+var parent = require('../../stable/typed-array/set-from-base64');
 require('../../modules/esnext.uint8-array.set-from-base64');
+
+module.exports = parent;
