+++ packages/core-js/actual/typed-array/set-from-hex.js
@@ -1,2 +1,5 @@
 'use strict';
+var parent = require('../../stable/typed-array/set-from-hex');
 require('../../modules/esnext.uint8-array.set-from-hex');
+
+module.exports = parent;
