+++ packages/core-js/actual/typed-array/to-base64.js
@@ -1,2 +1,5 @@
 'use strict';
+var parent = require('../../stable/typed-array/to-base64');
 require('../../modules/esnext.uint8-array.to-base64');
+
+module.exports = parent;
