+++ packages/core-js/actual/typed-array/to-hex.js
@@ -1,2 +1,5 @@
 'use strict';
+var parent = require('../../stable/typed-array/to-hex');
 require('../../modules/esnext.uint8-array.to-hex');
+
+module.exports = parent;
