+++ packages/core-js/es/typed-array/from-base64.js
@@ -1,0 +1,2 @@
+'use strict';
+require('../../modules/es.uint8-array.from-base64');
