+++ packages/core-js/es/typed-array/from-hex.js
@@ -1,0 +1,2 @@
+'use strict';
+require('../../modules/es.uint8-array.from-hex');
