+++ packages/core-js/es/typed-array/methods.js
@@ -32,3 +32,9 @@ require('../../modules/es.typed-array.to-reversed');
 require('../../modules/es.typed-array.to-sorted');
 require('../../modules/es.typed-array.with');
 require('../../modules/es.typed-array.iterator');
+require('../../modules/es.uint8-array.from-base64');
+require('../../modules/es.uint8-array.from-hex');
+require('../../modules/es.uint8-array.set-from-base64');
+require('../../modules/es.uint8-array.set-from-hex');
+require('../../modules/es.uint8-array.to-base64');
+require('../../modules/es.uint8-array.to-hex');
