+++ packages/core-js/es/typed-array/to-base64.js
@@ -1,0 +1,2 @@
+'use strict';
+require('../../modules/es.uint8-array.to-base64');
