+++ packages/core-js/stable/typed-array/from-hex.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/typed-array/from-hex');
+
+module.exports = parent;
