+++ packages/core-js/stable/typed-array/set-from-hex.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/typed-array/set-from-hex');
+
+module.exports = parent;
