+++ packages/core-js/stable/typed-array/to-base64.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/typed-array/to-base64');
+
+module.exports = parent;
