+++ packages/core-js/stable/typed-array/to-hex.js
@@ -1,0 +1,4 @@
+'use strict';
+var parent = require('../../es/typed-array/to-hex');
+
+module.exports = parent;
