+++ packages/core-js/stage/3.js
@@ -1,7 +1,6 @@
 'use strict';
 var parent = require('./4');
 
-require('../proposals/array-buffer-base64');
 require('../proposals/decorator-metadata-v2');
 require('../proposals/json-parse-with-source');
 require('../proposals/math-sum');
